New features        : Problems solved     : - FIXED: Improper integer type. Bugzilla-Bug-ID     : Additional comments :
git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@4874 4b12f95d-9313-0410-afc9-bd26427041f1

+++ kernel/Postprocessing/pprocerror.f90
@@ -2743,7 +2743,7 @@ contains
     real(DP), dimension(NDIM2D,TRIA_MAXNVE) :: Dcoord
     integer :: ncubp,ipoint
     integer(I32) :: celement
-    integer :: icoordSystem
+    integer(i32) :: icoordSystem
     real(DP) :: dlen,dpar1,dpar2
     
     ! Arrays for element distributions for every element
